Introduce feature flag to switch between new and old cache in SummaryConfirmationService - Keystore
