+++ conf/application.conf
@@ -168,3 +168,4 @@ company-officers-new-enabled = true
 isNewSummaryConfirmationCacheEnabled = false
 isNewSessionRepositoryCacheEnabled = false
 
+isMongoDBCacheEnabled = false
